Agregar pantalla de restablecimiento de contraseña: implementar la lógica para ingresar y confirmar la nueva contraseña, incluyendo validaciones y navegación.
